Squished branch: feat/sort-tsconfig-package
