+++ .changeset/sixty-mirrors-reply.md
@@ -1,0 +1,7 @@
+---
+"@tylerbu/sort-tsconfig": minor
+---
+
+Introducing sort-tsconfig, a CLI app to sort tsconfigs. It also provides a simple programmatic API.
+
+For more information, see the [sort-tsconfig readme](https://github.com/tylerbutler/tools-monorepo/blob/main/packages/sort-tsconfig/README.md).
