+++ fluidBuild.config.cjs
@@ -37,7 +37,7 @@ const config = {
 			dependsOn: ["compile"],
 		},
 		manifest: ["compile"],
-		readme: ["compile"],
+		readme: ["compile", "manifest"],
 		test: {
 			dependsOn: ["build"],
 		},
