+++ packages/cli-api/package.json
@@ -38,11 +38,8 @@
 		"chalk": "^5.3.0",
 		"cosmiconfig": "^9.0.0",
 		"debug": "^4.3.4",
-		"detect-indent": "^7.0.1",
 		"simple-git": "^3.24.0",
-		"sort-jsonc": "^1.0.1",
-		"strip-ansi": "^7.1.0",
-		"tiny-jsonc": "^1.0.1"
+		"strip-ansi": "^7.1.0"
 	},
 	"devDependencies": {
 		"@biomejs/biome": "1.8.3",
