+++ packages/cli-api/src/index.ts
@@ -16,10 +16,3 @@ export type {
 	LoggingFunction,
 	Logger,
 } from "./logger.js";
-export {
-	isSorted,
-	sortTsconfigFile,
-	type OrderList,
-	type SortTsconfigResult,
-	TsConfigSorter,
-} from "./tsconfig.js";
