+++ packages/sort-tsconfig/api-extractor.json
@@ -1,0 +1,4 @@
+{
+	"$schema": "https://developer.microsoft.com/json-schemas/api-extractor/v7/api-extractor.schema.json",
+	"extends": "../../config/api-extractor.base.json",
+}
