+++ packages/sort-tsconfig/src/index.ts
@@ -1,0 +1,8 @@
+// biome-ignore lint/performance/noBarrelFile:
+export {
+	isSorted,
+	sortTsconfigFile,
+	type OrderList,
+	type SortTsconfigResult,
+	TsConfigSorter,
+} from "./tsconfig.js";
