+++ packages/sort-tsconfig/test/data/tsconfig.sorted.json
