+++ packages/sort-tsconfig/test/data/tsconfig.unsorted.json
