+++ packages/sort-tsconfig/test/data/tsconfig.unsorted.unknown-keys.json
