+++ packages/sort-tsconfig/test/tsconfig.json
