+++ packages/sort-tsconfig/tsconfig.json
@@ -1,0 +1,11 @@
+{
+	"extends": "../../config/tsconfig.strict.json",
+	"include": ["src/**/*"],
+	"compilerOptions": {
+		"rootDir": "src",
+		"outDir": "esm",
+	},
+	"ts-node": {
+		"esm": true,
+	},
+}
